Fix coverage exclusions and organize code
