+++ app/src/test/java/io/capibaras/abcall/ui/viewmodels/NavigationViewModelTest.kt
@@ -1,13 +1,11 @@
-package io.capibaras.abcall
+package io.capibaras.abcall.ui.viewmodels
 
+import io.capibaras.abcall.R
 import io.capibaras.abcall.data.LogoutManager
 import io.capibaras.abcall.data.LogoutState
 import io.capibaras.abcall.data.TokenManager
 import io.capibaras.abcall.data.repositories.UsersRepository
-import io.capibaras.abcall.ui.viewmodels.ErrorUIState
-import io.capibaras.abcall.ui.viewmodels.SuccessUIState
-import io.capibaras.abcall.util.StateMediator
-import io.capibaras.abcall.viewmodels.NavigationViewModel
+import io.capibaras.abcall.ui.util.StateMediator
 import io.mockk.MockKAnnotations
 import io.mockk.Runs
 import io.mockk.coEvery
